chore(tests): migrate bq-card e2e tests to Stencil Vitest
